Multi auth praticamente completo, com paginas e controle de acesso

+++ resources/views/estagios/lista.blade.php
@@ -6,11 +6,12 @@
             <div class="col-md-8 col-md-offset-2">
                 <div class="panel panel-default">
                     <div class="panel-heading">
-                        @if (Auth::guard('aluno'))
+                        @if (Auth::getDefaultDriver() == 'aluno')
                             Meus Estágios
                             <a class="pull-right" href="{{ url('aluno/registrar-estagio') }}">Registrar Estágio</a>
                         @else
                             Estágios
+                            <a class="pull-right" href="{{ url('admin/registrar-estagio') }}">Registrar Estágio</a>
                         @endif
 
                     </div>
@@ -33,10 +34,17 @@
                                     <td>{{ $estagio->data_inicio }}</td>
                                     <td>{{ $estagio->data_termino }}</td>
                                     <td>
-                                        <a href="/aluno/estagios/{{ $estagio->id }}/editar" class="btn btn-default btn-sm">Editar</a>
-                                        {!! Form::open(['method' => 'DELETE', 'url' => 'aluno/estagios/'.$estagio->id, 'style' => 'display: inline;']) !!}
-                                        <button type="submit" class="btn btn-default btn-sm">Excluir</button>
-                                        {!! Form::close() !!}
+                                        @if (Auth::getDefaultDriver() == 'aluno')
+                                            <a href="/aluno/estagios/{{ $estagio->id }}/editar" class="btn btn-default btn-sm">Editar</a>
+                                            {!! Form::open(['method' => 'DELETE', 'url' => 'aluno/estagios/'.$estagio->id, 'style' => 'display: inline;']) !!}
+                                            <button type="submit" class="btn btn-default btn-sm">Excluir</button>
+                                            {!! Form::close() !!}
+                                        @else
+                                            <a href="/admin/estagios/{{ $estagio->id }}/editar" class="btn btn-default btn-sm">Editar</a>
+                                            {!! Form::open(['method' => 'DELETE', 'url' => 'admin/estagios/'.$estagio->id, 'style' => 'display: inline;']) !!}
+                                            <button type="submit" class="btn btn-default btn-sm">Excluir</button>
+                                            {!! Form::close() !!}
+                                        @endif
                                     </td>
                                 </tr>
                             @endforeach
